OctaHomeCore/requirements.txt
authy
django-js-reverse
django-polymorphic